refactor: Move shared project config out of routes
