+++ .prettierignore
@@ -1,3 +1,4 @@
 **/api
+**/build
 **/public
 pnpm-lock.yaml
