+++ app/routes/form.$version.docs.framework.$framework.examples.$.tsx
@@ -3,7 +3,7 @@ import type { LoaderFunctionArgs, MetaFunction } from '@remix-run/node'
 import { json } from '@remix-run/node'
 import { useLoaderData, useParams } from '@remix-run/react'
 import { DocTitle } from '~/components/DocTitle'
-import { repo, getBranch } from '~/routes/form'
+import { repo, getBranch } from '~/projects/form'
 import { seo } from '~/utils/seo'
 import { capitalize, slugToTitle } from '~/utils/utils'
 
@@ -13,7 +13,7 @@ export const loader = async (context: LoaderFunctionArgs) => {
   return json({ kind, name })
 }
 
-export const meta: MetaFunction = ({ data }) => {
+export const meta: MetaFunction<typeof loader> = ({ data }) => {
   return seo({
     title: `${capitalize(data.kind)} Form ${slugToTitle(
       data.name
