+++ app/routes/ranger.v1.docs.tsx
@@ -2,8 +2,7 @@ import * as React from 'react'
 import { FaDiscord, FaGithub } from 'react-icons/fa'
 import { Link } from '@remix-run/react'
 import type { MetaFunction } from '@remix-run/node'
-import { useRangerV1Config } from '~/routes/ranger.v1'
-import { gradientText } from '~/routes/ranger.v1._index'
+import { gradientText, useRangerV1Config } from '~/projects/ranger'
 import { seo } from '~/utils/seo'
 import type { DocsConfig } from '~/components/Docs'
 import { Docs } from '~/components/Docs'
