+++ app/routes/table.v8.tsx
@@ -2,11 +2,8 @@ import * as React from 'react'
 import { Link, Outlet, useLocation, useSearchParams } from '@remix-run/react'
 import { json } from '@remix-run/node'
 import { DefaultErrorBoundary } from '~/components/DefaultErrorBoundary'
-import type { DocsConfig } from '~/components/Docs'
 import { fetchRepoFile } from '~/utils/documents.server'
-import { useMatchesData } from '~/utils/utils'
-
-export const v8branch = 'main'
+import { v8branch } from '~/projects/table'
 
 export const loader = async () => {
   const config = await fetchRepoFile(
@@ -25,9 +22,6 @@ export const loader = async () => {
 }
 
 export const ErrorBoundary = DefaultErrorBoundary
-
-export const useReactTableV8Config = () =>
-  useMatchesData('/table/v8') as DocsConfig
 
 export default function RouteReactTable() {
   const [params] = useSearchParams()
