+++ app/routes/virtual.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from '@remix-run/react'
-import type { LoaderArgs, MetaFunction } from '@remix-run/node'
+import type { LoaderFunctionArgs, MetaFunction } from '@remix-run/node'
 import { redirect } from '@remix-run/node'
 import { seo } from '~/utils/seo'
 import { Scarf } from '~/components/Scarf'
@@ -14,7 +14,7 @@ export const meta: MetaFunction = () => {
   })
 }
 
-export const loader = async (context: LoaderArgs) => {
+export const loader = async (context: LoaderFunctionArgs) => {
   if (!context.request.url.includes('/virtual/v')) {
     return redirect(`${new URL(context.request.url).origin}/virtual/v3`)
   }
