+++ app/routes/virtual.v3.docs.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react'
 import { FaDiscord, FaGithub } from 'react-icons/fa'
 import { Link } from '@remix-run/react'
-import { useVirtualV3Config } from '~/routes/virtual.v3'
-import { gradientText } from '~/routes/virtual.v3._index'
+import { gradientText, useVirtualV3Config } from '~/projects/virtual'
 import { seo } from '~/utils/seo'
 import { Docs } from '~/components/Docs'
 import type { MetaFunction } from '@remix-run/react'
